Move GrabKitSources to GrabKitLib + various fixes
* Splitted the Model classes in categories : the implementations of
GRKAlbum and GRKPhoto classes allow read methods only. A "modify"
category has been added for each class, and is included only where
modification is needed, for classes allowed to modify the objects (i.e.
the grabbers)
There is no "modify" category for GRKImage yet, because there's no need
to modify GRKImage in the whole project. But GRKImage will follow this
design pattern in the future, if needed.

* added two properties to grabbers :
    * requiresConnection : returns YES for grabbers that needs to
"connect" ( through OAuth or other mechanisms). This property is *not*
related to the current state of the connection of the user, but is
related to the *need* of an authentication mechanism.
    * canLoadPagesDiscontinuously : returns YES if the grabber can load
pages of photos of an album in a discontinuous way ( load page 1, then
page 5, then page 0 ...)
For example, Instagram API doesn't offer pagination, so this property
on GRKInstagramGrabber returns NO.

* Performances improvements, by adding specific dispatch_async on the
proper queues where needed.

* Added preprocessor constants in GrabKit.h, to easily disable a
Grabber. see GRK_FACEBOOK_SERVICE, GRK_FLICKR_SERVICE, ...

*  Update in GRKConnectorsDispatcher for a better developper-experience
: the dispatcher is now adding itself as observer for specific
notifications, thus preventing to modify the appDelegate.

* Refined all the connectors

* Fix in GRKDeviceGrabber to load albums' covers